[REF] javascript to typescript
