+++ package.json
@@ -63,6 +63,10 @@
     "@babel/preset-react": "^7.22.3",
     "@jest/transform": "^29.6.2",
     "@tailwindcss/forms": "^0.5.3",
+    "@types/node": "^20.4.8",
+    "@types/react": "^18.2.18",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-pdf": "^7.0.0",
     "autoprefixer": "^10.4.14",
     "babel-jest": "^29.6.1",
     "eslint": "^8.41.0",
@@ -74,7 +78,8 @@
     "jest-transform-stub": "^2.0.0",
     "msw": "^1.2.1",
     "postcss": "^8.4.23",
-    "tailwindcss": "^3.3.2"
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.6"
   },
   "jest": {
     "transform": {
