--- src/utils/formartters/capitalizeWords.js
@@ -1,5 +1,0 @@
-export const capitalizeWords = (str) => {
-  return str.replace(/\w\S*/g, (txt) => {
-    return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-  });
-};
