+++ src/utils/formartters/capitalizeWords.ts
@@ -1,0 +1,5 @@
+export const capitalizeWords = (str: string) => {
+  return str.replace(/\w\S*/g, (txt) => {
+    return txt.charAt(0).toUpperCase() + txt.substring(1).toLowerCase();
+  });
+};
