--- src/utils/hooks/useDebounce.js
@@ -1,12 +1,0 @@
-import React, { useState, useEffect } from 'react';
-
-export const useDebounce = (value, delay = 500) => {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-    return () => clearTimeout(timeout);
-  }, [value, delay]);
-  return debouncedValue;
-};
